Enable timeline semaphores

+++ labs/enger/src/Renderer.h
@@ -23,13 +23,14 @@ namespace enger
         SwapChain& m_SwapChain;
 
         uint32_t m_CurrentFrame = 0;
-        uint32_t m_FrameNumber = 0;
+        uint64_t m_FrameNumber = 0;
 
         std::array<vk::UniqueCommandPool, FRAMES_IN_FLIGHT> m_CommandPools;
         // the cmdbuf will be destroyed when its parent pool is destroyed, so it doesn't need to be unique
         std::array<vk::CommandBuffer, FRAMES_IN_FLIGHT> m_CommandBuffers;
 
         std::array<vk::UniqueSemaphore, FRAMES_IN_FLIGHT> m_ImageAvailableSemaphores;
+        // this has as many elements as there are swapchain images
         std::vector<vk::UniqueSemaphore> m_RenderFinishedSemaphores;
         std::array<vk::UniqueFence, FRAMES_IN_FLIGHT> m_InFlightFences;
     };
